+++ index.js
@@ -21,6 +21,12 @@ const CryptoProProvider = () => {
   // algorithm GOST R 34.11-94.
   const CADESCOM_HASH_ALGORITHM_CP_GOST_3411 = 100;
 
+  // finding certificates from all storages
+  const CADESCOM_CONTAINER_STORE = 100;
+
+  // time of signing
+  const CAPICOM_AUTHENTICATED_ATTRIBUTE_SIGNING_TIME = 0;
+
   // provide access to cadesplugin_api
   const cadesplugin = window.cadesplugin;
 
@@ -46,7 +52,7 @@ const CryptoProProvider = () => {
 
       try {
         const store = cadesplugin.CreateObject("CAPICOM.Store");
-        store.Open();
+        store.Open(CADESCOM_CONTAINER_STORE);
 
         const certificates = store.Certificates;
         const count = certificates.Count;
@@ -93,7 +99,7 @@ const CryptoProProvider = () => {
 
         try {
           const store = yield cadesplugin.CreateObjectAsync("CAPICOM.Store");
-          yield store.Open();
+          yield store.Open(CADESCOM_CONTAINER_STORE);
 
           const certificates = yield store.Certificates;
           const count = yield certificates.Count;
@@ -148,9 +154,15 @@ const CryptoProProvider = () => {
                               .Find(CAPICOM_CERTIFICATE_FIND_SHA1_HASH, thumbprint)
                               .Item(1);
 
-
         const signer = cadesplugin.CreateObject("CAdESCOM.CPSigner");
         signer.Certificate = certificate;
+
+        const signingTimeAttr = cadesplugin.CreateObject("CADESCOM.CPAttribute");
+        signingTimeAttr.Name = CAPICOM_AUTHENTICATED_ATTRIBUTE_SIGNING_TIME;
+        signingTimeAttr.Value = _convertDate();
+
+        signer.AuthenticatedAttributes2.Add(signingTimeAttr);
+
 
         const signedData = cadesplugin.CreateObject("CAdESCOM.CadesSignedData");
         signedData.ContentEncoding = CADESCOM_BASE64_TO_BINARY;
@@ -193,9 +205,18 @@ const CryptoProProvider = () => {
           const certificate = yield certificates.Item(1);
 
           const signer = yield cadesplugin.CreateObjectAsync("CAdESCOM.CPSigner");
+
+          // Атрибут времени
+          const signingTimeAttr = yield cadesplugin.CreateObjectAsync("CADESCOM.CPAttribute");
+          yield signingTimeAttr.propset_Name(cadesplugin.CAPICOM_AUTHENTICATED_ATTRIBUTE_SIGNING_TIME);
+          yield signingTimeAttr.propset_Value(_convertDate());
+          const attr = yield signer.AuthenticatedAttributes2;
+          yield attr.Add(signingTimeAttr);
+
           yield signer.propset_Certificate(certificate);
 
           const signedData = yield cadesplugin.CreateObjectAsync("CAdESCOM.CadesSignedData");
+          yield signedData.propset_ContentEncoding(CADESCOM_BASE64_TO_BINARY);
           yield signedData.propset_Content(args[1]);
 
           const signature = yield signedData.SignCades(signer, CADESCOM_CADES_BES, true);
@@ -341,6 +362,18 @@ const CryptoProProvider = () => {
     }
 
     return window.btoa(str);
+  }
+
+  /**
+   * @function
+   * @name _convertDate
+   * @description Method returns date
+   * @return {string} date
+   */
+  const _convertDate = () => {
+    const date = new Date();
+    const navigator_name = navigator.appName;
+    return navigator_name == 'Microsoft Internet Explorer' ? date.getVarDate() : date
   }
 
   return {